Remove non agnostic timeslots from agnostic application
