+++ packages/technical-features/application/agnostic/jest.config.js
@@ -1,2 +1,2 @@
 module.exports =
-  require("@k8slens/jest").monorepoPackageConfig(__dirname).configForReact;
+  require("@k8slens/jest").monorepoPackageConfig(__dirname).configForNode;
