+++ packages/technical-features/application/agnostic/src/start-application/start-application.injectable.ts
@@ -1,21 +1,8 @@
 import { getInjectable, getInjectionToken } from "@ogre-tools/injectable";
-import { runManySyncFor, runManyFor } from "@ogre-tools/injectable-utils";
-import { beforeAnythingInjectionToken } from "./timeslots/before-anything-injection-token";
-import {
-  afterBeforeAnythingInjectionToken
-} from "./timeslots/after-before-anything-injection-token";
-import { untilReadyToStartInjectionToken } from "./triggers/until-ready-to-start-injection-token";
-import {
-  beforeApplicationIsLoadingInjectionToken
-} from "./timeslots/before-application-is-loading-injection-token";
-import { untilApplicationIsReadyToLoadInjectionToken } from "./triggers/until-application-is-ready-to-load-injection-token";
-import {
-  onLoadOfApplicationInjectionToken
-} from "./timeslots/on-load-of-application-injection-token";
-import {
-  afterApplicationIsLoadedInjectionToken
-} from "./timeslots/after-application-is-loaded-injection-token";
-import { untilApplicationIsShownInjectionToken } from "./triggers/until-application-is-shown-injection-token";
+import { runManyFor } from "@ogre-tools/injectable-utils";
+import { beforeApplicationIsLoadingInjectionToken } from "./timeslots/before-application-is-loading-injection-token";
+import { onLoadOfApplicationInjectionToken } from "./timeslots/on-load-of-application-injection-token";
+import { afterApplicationIsLoadedInjectionToken } from "./timeslots/after-application-is-loaded-injection-token";
 
 export type StartApplication = () => void;
 
@@ -28,20 +15,7 @@ const startApplicationInjectable = getInjectable({
   id: "start-application",
 
   instantiate: (di): StartApplication => {
-    const untilReadyToStart = di.inject(untilReadyToStartInjectionToken);
-    const untilApplicationIsReadyToLoad = di.inject(untilApplicationIsReadyToLoadInjectionToken);
-    const untilApplicationIsShown = di.inject(untilApplicationIsShownInjectionToken);
-
     const runManyAsync = runManyFor(di)
-    const runManySync = runManySyncFor(di)
-
-    const beforeAnything = runManySync(
-      beforeAnythingInjectionToken
-    );
-
-    const afterBeforeAnything = runManySync(
-      afterBeforeAnythingInjectionToken
-    );
 
     const beforeApplicationIsLoading = runManyAsync(
       beforeApplicationIsLoadingInjectionToken
@@ -54,16 +28,9 @@ const startApplicationInjectable = getInjectable({
     );
 
     return async () => {
-      beforeAnything();
-      afterBeforeAnything();
-
-      await untilReadyToStart()
-
       await beforeApplicationIsLoading();
-      await untilApplicationIsReadyToLoad();
 
       await onLoadOfApplication();
-      await untilApplicationIsShown();
 
       await afterApplicationIsLoaded();
     };
