--- packages/technical-features/application/agnostic/src/start-application/timeslots/after-before-anything-injection-token.ts
@@ -1,6 +1,0 @@
-import { getInjectionToken } from "@ogre-tools/injectable";
-import type { Runnable } from "@ogre-tools/injectable-utils";
-
-export const afterBeforeAnythingInjectionToken = getInjectionToken<Runnable>({
-  id: "after-before-anything",
-});
