--- packages/technical-features/application/agnostic/src/start-application/triggers/until-application-is-ready-to-load-injection-token.ts
@@ -1,8 +1,0 @@
-import { getInjectionToken } from "@ogre-tools/injectable";
-
-export type UntilApplicationIsReadyToLoad = () => Promise<void>;
-
-export const untilApplicationIsReadyToLoadInjectionToken =
-  getInjectionToken<UntilApplicationIsReadyToLoad>({
-    id: "until-application-is-ready-to-load-injection-token",
-  });
