--- packages/technical-features/application/agnostic/src/start-application/triggers/until-application-is-shown-injection-token.ts
@@ -1,8 +1,0 @@
-import { getInjectionToken } from "@ogre-tools/injectable";
-
-export type UntilApplicationIsShown = () => Promise<void>;
-
-export const untilApplicationIsShownInjectionToken =
-  getInjectionToken<UntilApplicationIsShown>({
-    id: "until-application-is-shown-injection-token",
-  });
