--- packages/technical-features/application/agnostic/src/start-application/triggers/until-ready-to-start-injection-token.ts
@@ -1,8 +1,0 @@
-import { getInjectionToken } from "@ogre-tools/injectable";
-
-export type UntilReadyToStart = () => Promise<void>;
-
-export const untilReadyToStartInjectionToken =
-  getInjectionToken<UntilReadyToStart>({
-    id: "until-ready-to-start-injection-token",
-  });
